+++ services/compute/mgmt/2021-12-01/compute/CHANGELOG.md
@@ -1,2 +1,14 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Structs
+
+1. UserArtifactSettings
+
+#### New Struct Fields
+
+1. GalleryApplicationVersionPublishingProfile.AdvancedSettings
+1. GalleryApplicationVersionPublishingProfile.Settings